Add files via upload
This python script creates a pandas dataframe with the data from the 'videos oct27-nov1.csv' file. It normalizes the numerical columns and one-hot-encodes the True/False columns. I converted the 'trending_dates' column to a true/false 'trending' columns (this could serve as the target variable). I wasn't sure how to deal with channel title, description, tags, data published, country, and channelId so I left them out for now. If the format of this script is weird/doesn't work with the other scripts let me know so I can adjust it :D